Allow zooming on CaptionedImage
